requirements.txt
grpcio
grpcio-tools
sqlalchemy
mysql-connector-python
yfinance
confluent-kafka
prometheus_client
pytz
